+++ Data/Browncoats.rte/Actors/Turrets/Thunderer/ThundererBase.lua
@@ -12,23 +12,9 @@ function Create(self)
 	end
 
 	self.AIMode = Actor.AIMODE_SENTRY;
-	
-	if self:GetNumberValue("KeepUnflipped") == -1 then
-		self.keepFlipped = true;
-	else
-		self.keepFlipped = false;
-	end
-	
-	if self.Turret then
-		if self.Turret.MountedDevice then
-			self.Turret.MountedDevice:SendMessage("SetKeepUnflipped", self:GetNumberValue("KeepUnflipped"));
-		end
-	end
 end
 
 function Update(self)
 	-- keep anything from moving us
 	self.Pos = self.pinPos;
-
-	self.HFlipped = self.keepFlipped;
 end